+++ src/io/flutter/run/test/FlutterTestLocationProvider.java
@@ -10,7 +10,6 @@ import com.intellij.execution.Location;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import com.jetbrains.lang.dart.psi.DartCallExpression;
-import io.flutter.util.DartTestLocationProviderZ;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
